don't question this. thank you.

+++ apis/game/level/block.go
@@ -8,6 +8,6 @@ type Block interface {
 	Chunk() Chunk
 	Level() Level
 
-	GetBlockType() int64
-	SetBlockType(value int64)
+	GetBlockType() int
+	SetBlockType(value int)
 }
